consolidate debug functions
